+++ gcloud/apigw/views/create_task.py
@@ -18,6 +18,7 @@ from django.views.decorators.csrf import csrf_exempt
 from django.views.decorators.http import require_POST
 
 from pipeline.exceptions import PipelineException
+from pipeline.core.constants import PE
 from pipeline_web.parser.validator import validate_web_pipeline_tree
 
 from blueapps.account.decorators import login_exempt
@@ -41,7 +42,35 @@ from gcloud.iam_auth.view_interceptors.apigw import CreateTaskInterceptor
 from gcloud.contrib.operate_record.decorators import record_operation
 from gcloud.contrib.operate_record.constants import RecordType, OperateType, OperateSource
 from packages.bkoauth.decorators import apigw_required
-from .utils import get_exclude_nodes_by_execute_nodes
+
+
+def get_pipeline_node(data):
+    """
+    递归获取当前实例中节点的 ID（不包括开始、结束节点、网关节点）
+    @param node_id_set: 节点 ID 集合
+    @param data: 流程数据
+    @return:
+    """
+    node_id_set = set()
+    for aid, act_data in list(data[PE.activities].items()):
+        node_id_set.add(aid)
+        if act_data[PE.type] == PE.SubProcess:
+            get_pipeline_node(node_id_set, act_data["pipeline"])
+    return list(node_id_set)
+
+
+def get_exclude_nodes_by_execute_nodes(execute_nodes, template):
+    """
+    @summary: 通过要选择执行的节点列表和任务模板获取要跳过执行的节点
+    @return: 要跳过执行的节点
+    """
+    pipeline_data = template.pipeline_tree
+    all_nodes = get_pipeline_node(pipeline_data)
+    # 排除掉在all_nodes中不存在的节点
+    execute_nodes = [node for node in execute_nodes if node in all_nodes]
+    # 筛选出在all_nodes中，不在execute_nodes的节点
+    exclude_nodes = [node for node in all_nodes if node not in execute_nodes]
+    return exclude_nodes
 
 
 @login_exempt
